+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 [Unreleased](https://github.com/FreddieChopin/bleeding-edge-toolchain/compare/190223...HEAD)
 -------------------------------------------------------------------------------------------------------
 
+### Added
+
+- `--quiet-build` option, which will make the build quiet(er).
+
 ### Changed
 
 - Updated gdb to version 8.3.0.